sample/astrogre: --bench-file mode + AOT-favourable bench harness
Three changes that together make the AOT specialisation gain
visible at the bench level:

1. VALUE typedef widened to int64_t (was int).  The framework's
   SPECIALIZE emits `(VALUE)<n>ULL` for uint64_t fields (e.g. class
   bitmaps), so a 32-bit VALUE silently truncated bm0..bm3 in baked
   SDs — most class members stopped matching after AOT.  Symptom
   was self-tests passing in interp but baked patterns returning
   nil.  Other ASTro samples already use int64_t for the same
   reason.

2. INIT() resolves the source directory via /proc/self/exe rather
   than passing "." to astro_cs_init.  When astrogre runs from a
   different cwd (e.g. bench/), "." pointed at the cwd and the
   generated SD .c files referenced node.h with the wrong absolute
   path, so make failed.  /proc/self/exe gives a stable bin-dir
   path regardless of cwd.

3. New `astrogre --bench-file FILE PATTERN ITERS [--aot|--plain]`
   mode that reads the file once and runs astrogre_search N times
   over the whole buffer.  Isolates the engine cost from the grep
   CLI's per-line getline / CTX-init overhead.  Plus
   bench/aot_bench.sh — a curated set of patterns chosen so the
   prefilter does NOT fire and the chain is long enough that
   dispatch is a meaningful share of wall time.

Result, 118 MB corpus, best-of-3 ms/iter (whole-buffer search):

  /(QQQ|RRR)+\d+/                  interp 2458   aot  950   2.59x
  /(QQQX|RRRX|SSSX)+/              interp 3139   aot 1051   2.99x
  /[a-z]\d[A-Z]\d[a-z]\d[A-Z]\d[a-z]/  interp  870  aot 396   2.20x
  /[a-z][0-9][a-z][0-9][a-z]/      interp 2.09   aot 1.07   1.96x
  /(\d+\.\d+\.\d+\.\d+)/           interp 36     aot 25     1.42x

The shape is consistent: AOT delivers a clean 2-3x when the chain
is long (alt + rep + multiple classes / literals), no prefilter
fires, and the search has to walk the whole input.  Single-class
patterns like /[A-Z]{50,}/ stay at 1.05x because the chain is
short enough that dispatch was already cheap.

+++ sample/astrogre/context.h
@@ -18,9 +18,11 @@
 #define ASTROGRE_MAX_GROUPS 32
 
 /* Match-result type returned by EVAL.  1 = continuation succeeded,
- * 0 = failed.  We use plain int rather than bool so that the framework's
- * VALUE-returning EVAL signature can carry it through. */
-typedef int VALUE;
+ * 0 = failed.  Width matters: the framework's SPECIALIZE emits
+ * `(VALUE)<n>ULL` for uint64_t fields (class bitmaps), so VALUE must
+ * be at least 64 bits wide or those casts truncate.  int64_t is the
+ * convention across the other samples for the same reason. */
+typedef int64_t VALUE;
 
 typedef enum {
     AGRE_ENC_ASCII = 0,
